Add mixin modules to public API
