+++ src/fatsecret/__init__.py
@@ -1,14 +1,26 @@
 from .core import FatsecretCore
 from .errors import (ApplicationError, AuthenticationError, BaseFatsecretError,
                      GeneralError, ParameterError)
+from .exercises import ExercisesMixin
 from .fatsecret import Fatsecret
+from .foods import FoodsMixin
+from .meals import MealsMixin
+from .profile import ProfileMixin
+from .recipes import RecipesMixin
+from .weight import WeightMixin
 
 __all__ = [
     "ApplicationError",
     "AuthenticationError",
     "BaseFatsecretError",
+    "ExercisesMixin",
     "Fatsecret",
     "FatsecretCore",
+    "FoodsMixin",
     "GeneralError",
+    "MealsMixin",
     "ParameterError",
+    "ProfileMixin",
+    "RecipesMixin",
+    "WeightMixin",
 ]
